+++ mre/src/Materials/MRayTracerMat.h
@@ -7,17 +7,17 @@ class MRayTracerMat : public Material {
 public:
 	MRayTracerMat();
 	MRayTracerMat(Vec3 c);
-	MRayTracerMat(Vec3 c, double i, int t);
+	MRayTracerMat(Vec3 c, float i, int t);
 
 	Vec3 getColor();
-	double getIOR();
+	float getIOR();
 	int getType();
 
 	std::string toString();
 
 private:
 	Vec3 color;
-	double IOR;
+	float IOR;
 	int type; //0 diffuse, 1 pure specular, 2 glass
